7-Setup-Login-MessageBar-and-SignIn-State: Refactor.
