+++ app/src/main/java/com/islaharper/dawnofandroid/di/RepositoryModule.kt
@@ -9,10 +9,10 @@ import com.islaharper.dawnofandroid.data.repository.DataStoreOperationsImpl
 import com.islaharper.dawnofandroid.data.repository.Repository
 import com.islaharper.dawnofandroid.domain.repository.DataStoreOperations
 import com.islaharper.dawnofandroid.domain.use_cases.UseCases
+import com.islaharper.dawnofandroid.domain.use_cases.readSignedInState.ReadSignedInStateUseCase
 import com.islaharper.dawnofandroid.domain.use_cases.read_onboarding.ReadOnBoardingStateUseCase
-import com.islaharper.dawnofandroid.domain.use_cases.read_signed_in_state.ReadSignedInStateUseCase
+import com.islaharper.dawnofandroid.domain.use_cases.saveSignedInState.SaveSignedInStateUseCase
 import com.islaharper.dawnofandroid.domain.use_cases.save_onboarding.SaveOnBoardingStateUseCase
-import com.islaharper.dawnofandroid.domain.use_cases.save_signed_in_state.SaveSignedInStateUseCase
 import com.islaharper.dawnofandroid.util.Constants.PREFS_NAME
 import dagger.Module
 import dagger.Provides
