+++ app/src/main/java/com/islaharper/dawnofandroid/domain/use_cases/UseCases.kt
@@ -1,9 +1,9 @@
 package com.islaharper.dawnofandroid.domain.use_cases
 
+import com.islaharper.dawnofandroid.domain.use_cases.readSignedInState.ReadSignedInStateUseCase
 import com.islaharper.dawnofandroid.domain.use_cases.read_onboarding.ReadOnBoardingStateUseCase
-import com.islaharper.dawnofandroid.domain.use_cases.read_signed_in_state.ReadSignedInStateUseCase
+import com.islaharper.dawnofandroid.domain.use_cases.saveSignedInState.SaveSignedInStateUseCase
 import com.islaharper.dawnofandroid.domain.use_cases.save_onboarding.SaveOnBoardingStateUseCase
-import com.islaharper.dawnofandroid.domain.use_cases.save_signed_in_state.SaveSignedInStateUseCase
 
 data class UseCases(
     val saveOnBoardingUseCase: SaveOnBoardingStateUseCase,
